+++ R/zzz.R
@@ -29,7 +29,15 @@
          envir = asNamespace("AMR"))
 }
 
-# maybe add survey later: "https://www.surveymonkey.com/r/AMR_for_R"
+.onAttach <- function(...) {
+  if (!interactive() || stats::runif(1) > 0.25 || isTRUE(as.logical(Sys.getenv("AMR_silentstart", FALSE)))) {
+    return()
+  }
+  packageStartupMessage("Thank you for using the AMR package! ",
+                        "If you have a minute, please anonymously fill in this short questionnaire to improve the package and its functionalities:",
+                        "\nhttps://msberends.github.io/AMR/survey.html",
+                        "\n[ permanently turn this message off with: Sys.setenv(AMR_silentstart = TRUE) ]")
+}
 
 create_MO_lookup <- function() {
   MO_lookup <- AMR::microorganisms